split publish and save method
