+++ save_file/main.py
@@ -1,0 +1,12 @@
+from Maya.save_file.source.ui import UI
+from Maya.save_file.source.controller import Controller
+
+
+def main():
+
+    UI_INSTANCE = Controller(UI)
+    # UI_INSTANCE.show()
+
+
+if __name__ == '__main__':
+    main()
