+++ save_file/source/core.py
@@ -64,15 +64,41 @@ def next_version(file_):
         raise ValueError(e)
 
 
-def publish(filepath):
-    path, name = os.path.split(filepath)
+# def publish(filepath):
+#     path, name = os.path.split(filepath)
+#
+#     publish_path = path.rsplit("/", 1)[0]
+#     publish_path = concat(publish_path, "PUBLISH", separator="/")
+#
+#     name, ext = os.path.splitext(name)
+#     publish_name = name.rsplit("_", 1)[0] + ext
+#
+#     publish = concat(publish_path, publish_name, separator="/")
+#
+#     copyfile(filepath, publish)
 
-    publish_path = path.rsplit("/", 1)[0]
-    publish_path = concat(publish_path, "PUBLISH", separator="/")
 
-    name, ext = os.path.splitext(name)
-    publish_name = name.rsplit("_", 1)[0] + ext
+def first_save(type, name, task):
+    """
 
-    publish = concat(publish_path, publish_name, separator="/")
+    Args:
+        type (str): chara, props, set
+        name (str): name of the asset
+        task (str): departement of the file: MOD, RIG, SHD, ANIM
 
-    copyfile(filepath, publish)
+    Returns:
+        str, str: versionned and published filepath
+
+    """
+    filename = concat(name, task, "001.ma", separator="_")
+    filepath = concat(PFE_PATH, "DATA/LIB", type, name, task, "SCENE/OLD", filename, separator="/")
+
+    save_as(filepath)
+
+    return filepath
+
+
+def clean_mode():
+    select_list_object()
+    freeze_transforms()
+    delete_history()
